+++ examples/doc_model_loadmodel.py
@@ -2,10 +2,12 @@
 #<examples/doc_model_loadmodel.py>
 import numpy as np
 import matplotlib.pyplot as plt
-from lmfit.model import Model, save_model, load_model
+
+from lmfit.model import load_model
 
 def mysine(x, amp, freq, shift):
     return amp * np.sin(x*freq + shift)
+
 
 data = np.loadtxt('sinedata.dat')
 x = data[:, 0]
@@ -17,10 +19,10 @@ params['shift'].max = 1
 params['shift'].min = -1
 params['amp'].min = 0.0
 
-result = model.fit(y,  params, x=x)
+result = model.fit(y, params, x=x)
 print(result.fit_report())
 
-plt.plot(x, y,         'bo')
+plt.plot(x, y, 'bo')
 plt.plot(x, result.best_fit, 'r-')
 plt.show()