+++ examples/doc_model_savemodelresult2.py
@@ -14,17 +14,17 @@ pars = exp_mod.guess(y, x=x)
 gauss1 = GaussianModel(prefix='g1_')
 pars.update(gauss1.make_params())
 
-pars['g1_center'].set(105, min=75, max=125)
-pars['g1_sigma'].set(15, min=3)
-pars['g1_amplitude'].set(2000, min=10)
+pars['g1_center'].set(value=105, min=75, max=125)
+pars['g1_sigma'].set(value=15, min=3)
+pars['g1_amplitude'].set(value=2000, min=10)
 
 gauss2 = GaussianModel(prefix='g2_')
 
 pars.update(gauss2.make_params())
 
-pars['g2_center'].set(155, min=125, max=175)
-pars['g2_sigma'].set(15, min=3)
-pars['g2_amplitude'].set(2000, min=10)
+pars['g2_center'].set(value=155, min=125, max=175)
+pars['g2_sigma'].set(value=15, min=3)
+pars['g2_amplitude'].set(value=2000, min=10)
 
 mod = gauss1 + gauss2 + exp_mod
 
@@ -35,4 +35,5 @@ result = mod.fit(y, pars, x=x)
 save_modelresult(result, 'nistgauss_modelresult.sav')
 
 print(result.fit_report())
+
 #<end examples/doc_model_savemodelresult2.py>