+++ lmfit/models.py
@@ -5,10 +5,10 @@ from . import lineshapes
 from .asteval import Interpreter
 from .astutils import get_ast_names
 from .lineshapes import (breit_wigner, damped_oscillator, dho, donaich,
-                         expgaussian, exponential, gaussian, linear, logistic,
-                         lognormal, lorentzian, moffat, parabolic, pearson7,
-                         powerlaw, pvoigt, rectangle, skewed_gaussian,
-                         skewed_voigt, step, students_t, voigt)
+                         expgaussian, exponential, gaussian, linear, lognormal,
+                         lorentzian, moffat, parabolic, pearson7, powerlaw,
+                         pvoigt, rectangle, skewed_gaussian, step, students_t,
+                         voigt)
 from .model import Model
 
 
@@ -155,6 +155,7 @@ class ConstantModel(Model):
     __init__.__doc__ = COMMON_INIT_DOC
     guess.__doc__ = COMMON_GUESS_DOC
 
+
 class ComplexConstantModel(Model):
     """Complex constant model, with wo Parameters: ``re``, and ``im``.
 
@@ -166,7 +167,7 @@ class ComplexConstantModel(Model):
     """
 
     def __init__(self, independent_vars=['x'], prefix='', nan_policy='raise',
-                 name=None,  **kwargs):
+                 name=None, **kwargs):
         kwargs.update({'prefix': prefix, 'nan_policy': nan_policy,
                        'independent_vars': independent_vars})
 
@@ -291,7 +292,7 @@ class PolynomialModel(Model):
 
 class GaussianModel(Model):
     r"""A model based on a Gaussian or normal distribution lineshape (see
-    http://en.wikipedia.org/wiki/Normal_distribution), with three Parameters:
+    https://en.wikipedia.org/wiki/Normal_distribution), with three Parameters:
     ``amplitude``, ``center``, and ``sigma``.
     In addition, parameters ``fwhm`` and ``height`` are included as constraints
     to report full width at half maximum and maximum peak height, respectively.
@@ -329,7 +330,7 @@ class GaussianModel(Model):
 
 class LorentzianModel(Model):
     r"""A model based on a Lorentzian or Cauchy-Lorentz distribution function
-    (see http://en.wikipedia.org/wiki/Cauchy_distribution), with three Parameters:
+    (see https://en.wikipedia.org/wiki/Cauchy_distribution), with three Parameters:
     ``amplitude``, ``center``, and ``sigma``.
     In addition, parameters ``fwhm`` and ``height`` are included as constraints
     to report full width at half maximum and maximum peak height, respectively.
@@ -366,7 +367,7 @@ class LorentzianModel(Model):
 
 class VoigtModel(Model):
     r"""A model based on a Voigt distribution function (see
-    http://en.wikipedia.org/wiki/Voigt_profile), with four Parameters:
+    https://en.wikipedia.org/wiki/Voigt_profile), with four Parameters:
     ``amplitude``, ``center``, ``sigma``, and ``gamma``.  By default,
     ``gamma`` is constrained to have a value equal to ``sigma``, though it
     can be varied independently.  In addition, parameters ``fwhm`` and
@@ -423,7 +424,7 @@ class VoigtModel(Model):
 
 class PseudoVoigtModel(Model):
     r"""A model based on a pseudo-Voigt distribution function
-    (see http://en.wikipedia.org/wiki/Voigt_profile#Pseudo-Voigt_Approximation),
+    (see https://en.wikipedia.org/wiki/Voigt_profile#Pseudo-Voigt_Approximation),
     which is a weighted sum of a Gaussian and Lorentzian distribution function
     that share values for ``amplitude`` (:math:`A`), ``center`` (:math:`\mu`)
     and full width at half maximum ``fwhm`` (and so have  constrained values of
@@ -451,7 +452,7 @@ class PseudoVoigtModel(Model):
                        'independent_vars': independent_vars})
         super(PseudoVoigtModel, self).__init__(pvoigt, **kwargs)
         self.set_param_hint('sigma', min=0)
-        self.set_param_hint('fraction', value=0.5)
+        self.set_param_hint('fraction', value=0.5, min=0.0, max=1.0)
         self.set_param_hint('fwhm', expr=fwhm_expr(self))
         fmt = ("(((1-{prefix:s}fraction)*{prefix:s}amplitude)/"
                "({prefix:s}sigma*sqrt(pi/log(2)))+"
@@ -461,7 +462,7 @@ class PseudoVoigtModel(Model):
 
     def guess(self, data, x=None, negative=False, **kwargs):
         pars = guess_from_peak(self, data, x, negative, ampscale=1.25)
-        pars['%sfraction' % self.prefix].set(value=0.5)
+        pars['%sfraction' % self.prefix].set(value=0.5, min=0.0, max=1.0)
         return update_param_vals(pars, self.prefix, **kwargs)
 
     __init__.__doc__ = COMMON_INIT_DOC
@@ -504,7 +505,7 @@ class MoffatModel(Model):
 
 class Pearson7Model(Model):
     r"""A model based on a Pearson VII distribution (see
-    http://en.wikipedia.org/wiki/Pearson_distribution#The_Pearson_type_VII_distribution),
+    https://en.wikipedia.org/wiki/Pearson_distribution#The_Pearson_type_VII_distribution),
     with four parameters: ``amplitude`` (:math:`A`), ``center``
     (:math:`\mu`), ``sigma`` (:math:`\sigma`), and ``exponent`` (:math:`m`) in
 
@@ -545,7 +546,7 @@ class Pearson7Model(Model):
 
 class StudentsTModel(Model):
     r"""A model based on a Student's t distribution function (see
-    http://en.wikipedia.org/wiki/Student%27s_t-distribution), with three Parameters:
+    https://en.wikipedia.org/wiki/Student%27s_t-distribution), with three Parameters:
     ``amplitude`` (:math:`A`), ``center`` (:math:`\mu`) and ``sigma`` (:math:`\sigma`) in
 
     .. math::
@@ -573,7 +574,7 @@ class StudentsTModel(Model):
 
 class BreitWignerModel(Model):
     r"""A model based on a Breit-Wigner-Fano function (see
-    http://en.wikipedia.org/wiki/Fano_resonance), with four Parameters:
+    https://en.wikipedia.org/wiki/Fano_resonance), with four Parameters:
     ``amplitude`` (:math:`A`), ``center`` (:math:`\mu`),
     ``sigma`` (:math:`\sigma`), and ``q`` (:math:`q`) in
 
@@ -600,7 +601,7 @@ class BreitWignerModel(Model):
 
 class LognormalModel(Model):
     r"""A model based on the Log-normal distribution function
-    (see http://en.wikipedia.org/wiki/Lognormal), with three Parameters
+    (see https://en.wikipedia.org/wiki/Lognormal), with three Parameters
     ``amplitude`` (:math:`A`), ``center`` (:math:`\mu`) and ``sigma``
     (:math:`\sigma`) in
 
@@ -627,7 +628,7 @@ class LognormalModel(Model):
 
 class DampedOscillatorModel(Model):
     r"""A model based on the Damped Harmonic Oscillator Amplitude
-    (see http://en.wikipedia.org/wiki/Harmonic_oscillator#Amplitude_part), with
+    (see https://en.wikipedia.org/wiki/Harmonic_oscillator#Amplitude_part), with
     three Parameters:  ``amplitude`` (:math:`A`), ``center`` (:math:`\mu`) and
     ``sigma`` (:math:`\sigma`) in
 
@@ -654,7 +655,7 @@ class DampedOscillatorModel(Model):
 
 class DampedHarmonicOscillatorModel(Model):
     r"""A model based on a variation of the Damped Harmonic Oscillator (see
-    http://en.wikipedia.org/wiki/Harmonic_oscillator), following the
+    https://en.wikipedia.org/wiki/Harmonic_oscillator), following the
     definition given in DAVE/PAN (see https://www.ncnr.nist.gov/dave/) with
     four Parameters: ``amplitude`` (:math:`A`), ``center`` (:math:`\mu`),
     ``sigma`` (:math:`\sigma`), and ``gamma`` (:math:`\gamma`) in
@@ -670,7 +671,7 @@ class DampedHarmonicOscillatorModel(Model):
                  **kwargs):
         kwargs.update({'prefix': prefix, 'nan_policy': nan_policy,
                        'independent_vars': independent_vars})
-        super(DampedHarmonicOscillatorModel, self).__init__(dho,  **kwargs)
+        super(DampedHarmonicOscillatorModel, self).__init__(dho, **kwargs)
 
     def guess(self, data, x=None, negative=False, **kwargs):
         pars = guess_from_peak(self, data, x, negative,
@@ -684,7 +685,7 @@ class DampedHarmonicOscillatorModel(Model):
 
 class ExponentialGaussianModel(Model):
     r"""A model of an Exponentially modified Gaussian distribution
-    (see http://en.wikipedia.org/wiki/Exponentially_modified_Gaussian_distribution) with
+    (see https://en.wikipedia.org/wiki/Exponentially_modified_Gaussian_distribution) with
     four Parameters ``amplitude`` (:math:`A`), ``center`` (:math:`\mu`),
     ``sigma`` (:math:`\sigma`), and  ``gamma`` (:math:`\gamma`) in
 
@@ -715,7 +716,7 @@ class ExponentialGaussianModel(Model):
 
 class SkewedGaussianModel(Model):
     r"""A variation of the Exponential Gaussian, this uses a skewed normal distribution
-    (see http://en.wikipedia.org/wiki/Skew_normal_distribution), with Parameters
+    (see https://en.wikipedia.org/wiki/Skew_normal_distribution), with Parameters
     ``amplitude`` (:math:`A`), ``center`` (:math:`\mu`),  ``sigma`` (:math:`\sigma`),
     and ``gamma`` (:math:`\gamma`) in
 
@@ -737,7 +738,7 @@ class SkewedGaussianModel(Model):
                  **kwargs):
         kwargs.update({'prefix': prefix, 'nan_policy': nan_policy,
                        'independent_vars': independent_vars})
-        super(SkewedGaussianModel, self).__init__(skewed_gaussian,  **kwargs)
+        super(SkewedGaussianModel, self).__init__(skewed_gaussian, **kwargs)
         self.set_param_hint('sigma', min=0)
 
     def guess(self, data, x=None, negative=False, **kwargs):
@@ -750,7 +751,7 @@ class SkewedGaussianModel(Model):
 
 class DonaichModel(Model):
     r"""A model of an Doniach Sunjic asymmetric lineshape
-    (see http://www.casaxps.com/help_manual/line_shapes.htm), used in
+    (see https://www.casaxps.com/help_manual/line_shapes.htm), used in
     photo-emission, with four Parameters ``amplitude`` (:math:`A`),
     ``center`` (:math:`\mu`), ``sigma`` (:math:`\sigma`), and ``gamma``
     (:math:`\gamma`) in
@@ -766,7 +767,7 @@ class DonaichModel(Model):
                  **kwargs):
         kwargs.update({'prefix': prefix, 'nan_policy': nan_policy,
                        'independent_vars': independent_vars})
-        super(DonaichModel, self).__init__(donaich,  **kwargs)
+        super(DonaichModel, self).__init__(donaich, **kwargs)
 
     def guess(self, data, x=None, negative=False, **kwargs):
         pars = guess_from_peak(self, data, x, negative, ampscale=0.5)
@@ -777,7 +778,7 @@ class DonaichModel(Model):
 
 
 class PowerLawModel(Model):
-    r"""A model based on a Power Law (see http://en.wikipedia.org/wiki/Power_law),
+    r"""A model based on a Power Law (see https://en.wikipedia.org/wiki/Power_law),
     with two Parameters: ``amplitude`` (:math:`A`), and ``exponent`` (:math:`k`), in:
 
     .. math::
@@ -795,7 +796,7 @@ class PowerLawModel(Model):
     def guess(self, data, x=None, **kwargs):
         try:
             expon, amp = np.polyfit(np.log(x+1.e-14), np.log(data+1.e-14), 1)
-        except:
+        except TypeError:
             expon, amp = 1, np.log(abs(max(data)+1.e-9))
 
         pars = self.make_params(amplitude=np.exp(amp), exponent=expon)
@@ -807,7 +808,7 @@ class PowerLawModel(Model):
 
 class ExponentialModel(Model):
     r"""A model based on an exponential decay function
-    (see http://en.wikipedia.org/wiki/Exponential_decay) with two Parameters:
+    (see https://en.wikipedia.org/wiki/Exponential_decay) with two Parameters:
     ``amplitude`` (:math:`A`), and ``decay`` (:math:`\tau`), in:
 
     .. math::
@@ -823,10 +824,9 @@ class ExponentialModel(Model):
         super(ExponentialModel, self).__init__(exponential, **kwargs)
 
     def guess(self, data, x=None, **kwargs):
-
         try:
             sval, oval = np.polyfit(x, np.log(abs(data)+1.e-15), 1)
-        except:
+        except TypeError:
             sval, oval = 1., np.log(abs(max(data)+1.e-9))
         pars = self.make_params(amplitude=np.exp(oval), decay=-1.0/sval)
         return update_param_vals(pars, self.prefix, **kwargs)
@@ -844,7 +844,7 @@ class StepModel(Model):
     - ``linear`` (the default)
     - ``atan`` or ``arctan`` for an arc-tangent function
     - ``erf`` for an error function
-    - ``logistic`` for a logistic function (see http://en.wikipedia.org/wiki/Logistic_function)
+    - ``logistic`` for a logistic function (see https://en.wikipedia.org/wiki/Logistic_function)
 
     The step function starts with a value 0, and ends with a value of
     :math:`A` rising to :math:`A/2` at :math:`\mu`, with :math:`\sigma`
@@ -896,7 +896,7 @@ class RectangleModel(Model):
     - ``linear`` (the default)
     - ``atan`` or ``arctan`` for an arc-tangent function
     - ``erf`` for an error function
-    - ``logistic`` for a logistic function (see http://en.wikipedia.org/wiki/Logistic_function)
+    - ``logistic`` for a logistic function (see https://en.wikipedia.org/wiki/Logistic_function)
 
     The function starts with a value 0, transitions to a value of
     :math:`A`, taking the value :math:`A/2` at :math:`\mu_1`, with :math:`\sigma_1`